+++ src/tooling/piral-cli/src/common/constants.ts
@@ -1,5 +1,6 @@
 export const defaultRegistry = 'https://registry.npmjs.org/';
 export const filesTar = 'files';
+export const packageJson = 'package.json';
 export const filesOnceTar = 'files_once';
 export const piralBaseRoot = 'piral-base/package.json';
 export const frameworkLibs = ['piral-native' as const, 'piral' as const, 'piral-core' as const, 'piral-base' as const];
@@ -11,6 +12,7 @@ export const bundlerNames = [
   'parcel' as const,
   'parcel2' as const,
   'rollup' as const,
+  'rspack' as const,
   'webpack' as const,
   'webpack5' as const,
   'vite' as const,