+++ src/tooling/piral-cli/src/common/utils.ts
@@ -1,0 +1,3 @@
+export function onlyUnique<T>(value: T, index: number, self: Array<T>) {
+  return self.indexOf(value) === index;
+}