+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Piral Changelog
 
+## 0.15.9 (tbd)
+
+- Fixed transport of `state` in routes with `piral-blazor`
+- Fixed issue with `piral-svelte/extend-webpack` helper
+- Fixed treatment of npm aliased packages as shared dependencies (#593)
+- Fixed Node.js engine constraint in `piral-cli` to "14.18.0"
+- Changed `pilet pack` to include specified `files` and the *README.md* if available
+- Added convenience module `piral-vue-3/extend-webpack`
+- Added warnings when plugins are included in the `importmap` (#591)
+- Added options to configure default debug flags for Piral instances (#590)
+- Added more properties and attributes to `piral-extension` matching `ExtensionSlot`
+
 ## 0.15.8 (March 8, 2023)
 
 - Fixed value of `schemaVersion` in *pilet.json* being ignored (#585)